Halve what a search over the cells of a continent costs
A query over the overlay spent its time in three places that had little
to do with searching: finding the table of a cell, working out which
level to step over, and asking the partition about the far end of every
arc it looked at. Each is answered here the way OSRM answers it.

The tables of a level sit in a vector indexed by cell rather than in a
map behind a lock, so reading one is a load and a branch. The cells a
node lies in on every level are packed into one word, so the level to
step over is the top bit of the difference of two words rather than a
walk of six arrays. The level at which each arc leaves its cell is
worked out once and held beside the arcs, so the search reads a byte in
step with them instead of jumping into an array as wide as the graph.
And the nodes are numbered so that the ones a coarse cell has on its
border come first, which puts everything such a search reads into the
first sixth of every array over the nodes.

An arc is now four bytes to a side rather than eight, which takes a
continent's adjacency array from six hundred and seventy five megabytes
to three hundred and thirty eight. What a node id is has not changed:
narrowing that alone buys nothing, as the padding puts the arc back at
sixteen bytes, and it costs a conversion at every place a node indexes
anything.

The numbering is a step of its own rather than something a query does as
it starts, so an instance is written under it once and read under it
after. What is already on disk still reads: the shape an arc has in a
file is pinned apart from the shape the graph holds it in.

Measured over eighteen million nodes and four thousand eight hundred
pairs drawn from every rank, a search from both ends went from eighty
four microseconds to forty seven at the median, and from eight hundred
and ninety nine to four hundred and eighty five at the mean. Every pair
comes back with the distance it came back with before.

+++ src/grid_graph.rs
@@ -48,7 +48,7 @@ pub fn node_at(side: usize, row: usize, column: usize) -> NodeID {
 /// different amount to cross in each direction. A query that assumed otherwise
 /// would be right on a grid of the other kind and wrong on a road.
 #[must_use]
-pub fn grid_edges(side: usize, both_ways: bool) -> Vec<InputEdge<usize>> {
+pub fn grid_edges(side: usize, both_ways: bool) -> Vec<InputEdge<u32>> {
     let mut edges = Vec::new();
     for row in 0..side {
         for column in 0..side {
@@ -142,7 +142,7 @@ pub fn grid_directory(side: usize) -> LevelDirectory {
 /// Panics unless the side is a power of two of at least four. See
 /// [`grid_directory`].
 #[must_use]
-pub fn grid(side: usize, both_ways: bool) -> (StaticGraph<usize>, LevelDirectory) {
+pub fn grid(side: usize, both_ways: bool) -> (StaticGraph<u32>, LevelDirectory) {
     (
         StaticGraph::new(grid_edges(side, both_ways)),
         grid_directory(side),
